draw static canvas with server state
